Fix screen utilization at 80x25
While food is scaled to account for variable room sizes due to variable
screen sizes, several other game mechanics (spell regeneration, rare events)
are not. It seems simplest to just not base the size of the level on the
size of the screen. This is set via ur_lines and ur_cols in main.c.